+++ core/src/main/scala/com/raphtory/api/analysis/table/TableImplementation.scala
@@ -1,8 +1,7 @@
 package com.raphtory.api.analysis.table
 
 import com.raphtory.api.output.sink.Sink
-import com.raphtory.api.querytracker.ProgressTracker
-import com.raphtory.api.querytracker.QueryProgressTracker
+import com.raphtory.api.querytracker.{ProgressTracker, QueryProgressTracker, TableOutputTracker}
 import com.raphtory.internals.components.output.TableOutputSink
 import com.raphtory.internals.components.querymanager.Query
 import com.raphtory.internals.management._
